+++ core/src/main/java/org/eqasim/core/components/config/EqasimConfigGroup.java
@@ -4,6 +4,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
+import org.eqasim.core.analysis.DistanceUnit;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigGroup;
 import org.matsim.core.config.ReflectiveConfigGroup;
@@ -11,11 +12,27 @@ import org.matsim.core.config.ReflectiveConfigGroup;
 public class EqasimConfigGroup extends ReflectiveConfigGroup {
 	public final static String GROUP_NAME = "eqasim";
 
-	private final static String CROSSING_PENALTY = "crossingPenalty";
 	private final static String SAMPLE_SIZE = "sampleSize";
+	private final static String DISTANCE_UNIT = "distanceUnit";
+
+	private final static String CROSSING_PENALTY = "crossingPenalty";
+
+	private final static String MODE_PARAMETERS_PATH = "modeParametersPath";
+	private final static String COST_PARAMETERS_PATH = "costParametersPath";
+
+	private final static String TRIP_ANALYSIS_INTERVAL = "tripAnalysisInterval";
+	private final static String TRIP_ANALYSIS_DISTANCE_UNIT = "tripAnalysisDistanceUnit";
+
+	private double sampleSize = 1.0;
+	private DistanceUnit distanceUnit = DistanceUnit.meter;
 
 	private double crossingPenalty = 3.0;
-	private double sampleSize = 1.0;
+
+	private String modeParametersPath = null;
+	private String costParametersPath = null;
+
+	private int tripAnalysisInterval = 0;
+	private DistanceUnit tripAnalysisDistanceUnit = DistanceUnit.meter;
 
 	public EqasimConfigGroup() {
 		super(GROUP_NAME);
@@ -155,5 +172,55 @@ public class EqasimConfigGroup extends ReflectiveConfigGroup {
 		}
 
 		return (EqasimConfigGroup) config.getModules().get(GROUP_NAME);
+	}
+
+	@StringGetter(MODE_PARAMETERS_PATH)
+	public String getModeParametersPath() {
+		return modeParametersPath;
+	}
+
+	@StringSetter(MODE_PARAMETERS_PATH)
+	public void setModeParametersPath(String modeParametersPath) {
+		this.modeParametersPath = modeParametersPath;
+	}
+
+	@StringGetter(COST_PARAMETERS_PATH)
+	public String getCostParametersPath() {
+		return costParametersPath;
+	}
+
+	@StringSetter(COST_PARAMETERS_PATH)
+	public void setCostParametersPath(String costParametersPath) {
+		this.costParametersPath = costParametersPath;
+	}
+
+	@StringGetter(TRIP_ANALYSIS_INTERVAL)
+	public int getTripAnalysisInterval() {
+		return tripAnalysisInterval;
+	}
+
+	@StringSetter(TRIP_ANALYSIS_INTERVAL)
+	public void setTripAnalysisInterval(int tripAnalysisInterval) {
+		this.tripAnalysisInterval = tripAnalysisInterval;
+	}
+
+	@StringGetter(DISTANCE_UNIT)
+	public DistanceUnit getDistanceUnit() {
+		return distanceUnit;
+	}
+
+	@StringSetter(DISTANCE_UNIT)
+	public void setDistanceUnit(DistanceUnit distanceUnit) {
+		this.distanceUnit = distanceUnit;
+	}
+
+	@StringGetter(TRIP_ANALYSIS_DISTANCE_UNIT)
+	public DistanceUnit getTripAnalysisDistanceUnit() {
+		return tripAnalysisDistanceUnit;
+	}
+
+	@StringSetter(TRIP_ANALYSIS_DISTANCE_UNIT)
+	public void setTripAnalysisDistanceUnit(DistanceUnit tripAnalysisDistanceUnit) {
+		this.tripAnalysisDistanceUnit = tripAnalysisDistanceUnit;
 	}
 }