+++ core/src/main/java/org/eqasim/core/simulation/EqasimConfigurator.java
@@ -3,12 +3,14 @@ package org.eqasim.core.simulation;
 import java.util.Arrays;
 import java.util.List;
 
+import org.eqasim.core.components.EqasimComponentsModule;
 import org.eqasim.core.components.config.EqasimConfigGroup;
 import org.eqasim.core.components.traffic.EqasimTrafficQSimModule;
 import org.eqasim.core.components.transit.EqasimTransitModule;
 import org.eqasim.core.components.transit.EqasimTransitQSimModule;
 import org.eqasim.core.components.transit.routing.DefaultEnrichedTransitRoute;
 import org.eqasim.core.components.transit.routing.DefaultEnrichedTransitRouteFactory;
+import org.eqasim.core.simulation.calibration.CalibrationConfigGroup;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.population.Person;
@@ -28,15 +30,16 @@ public class EqasimConfigurator {
 		return new ConfigGroup[] { //
 				new SwissRailRaptorConfigGroup(), //
 				new EqasimConfigGroup(), //
-				new DiscreteModeChoiceConfigGroup() //
-		};
+				new DiscreteModeChoiceConfigGroup(), //
+				new CalibrationConfigGroup() };
 	}
 
 	static public List<AbstractModule> getModules() {
 		return Arrays.asList( //
 				new SwissRailRaptorModule(), //
 				new EqasimTransitModule(), //
-				new DiscreteModeChoiceModule() //
+				new DiscreteModeChoiceModule(), //
+				new EqasimComponentsModule() //
 		);
 	}