+++ core/src/main/java/org/eqasim/core/simulation/analysis/EqasimAnalysisModule.java
@@ -1,0 +1,30 @@
+package org.eqasim.core.simulation.analysis;
+
+import org.eqasim.core.analysis.DefaultPersonAnalysisFilter;
+import org.eqasim.core.analysis.PersonAnalysisFilter;
+import org.eqasim.core.analysis.TripListener;
+import org.matsim.api.core.v01.network.Network;
+import org.matsim.core.config.groups.PlansCalcRouteConfigGroup;
+import org.matsim.core.controler.AbstractModule;
+import org.matsim.core.router.MainModeIdentifier;
+
+import com.google.inject.Provides;
+import com.google.inject.Singleton;
+
+public class EqasimAnalysisModule extends AbstractModule {
+	@Override
+	public void install() {
+		addControlerListenerBinding().to(AnalysisOutputListener.class);
+		bind(DefaultPersonAnalysisFilter.class);
+		bind(PersonAnalysisFilter.class).to(DefaultPersonAnalysisFilter.class);
+	}
+
+	@Provides
+	@Singleton
+	public TripListener provideTripListener(Network network, MainModeIdentifier mainModeIdentifier,
+			PlansCalcRouteConfigGroup routeConfig, PersonAnalysisFilter personFilter) {
+		return new TripListener(network, type -> type.endsWith("interaction"), mainModeIdentifier,
+				routeConfig.getNetworkModes(), personFilter);
+	}
+
+}