+++ core/src/main/java/org/eqasim/core/simulation/mode_choice/utilities/estimators/PtUtilityEstimator.java
@@ -45,12 +45,12 @@ public class PtUtilityEstimator implements UtilityEstimator {
 
 	protected double estimateMonetaryCostUtility(PtVariables variables) {
 		return parameters.betaCost_u_MU * EstimatorUtils.interaction(variables.euclideanDistance_km,
-				parameters.referenceEuclideanDistance_km, parameters.lambdaCostEuclideanDistance);
+				parameters.referenceEuclideanDistance_km, parameters.lambdaCostEuclideanDistance) * variables.cost_MU;
 	}
 
 	@Override
 	public double estimateUtility(Person person, DiscreteModeChoiceTrip trip, List<? extends PlanElement> elements) {
-		PtVariables variables = predictor.predict(person, trip, elements);
+		PtVariables variables = predictor.predictVariables(person, trip, elements);
 
 		double utility = 0.0;